+++ src/Core/Content/App/Lifecycle/Registration/AppHandshakeInterface.php
@@ -2,9 +2,11 @@
 
 namespace Swag\SaasConnect\Core\Content\App\Lifecycle\Registration;
 
+use Psr\Http\Message\RequestInterface;
+
 interface AppHandshakeInterface
 {
-    public function fetchUrl(): string;
+    public function assembleRequest(): RequestInterface;
 
     public function fetchAppProof(): string;
 }